Updated OpCert Generation in the Operational Credentials Issuer.
Problem: In the current implementation when ICAC and RCAC are
found in the storage the Subject DN for these certificates is generated
separately. This might be an issue if after software update the way
Subject DN is generated is changed.

Fix: The proper implementation is to extract Subject DN from ICAC and RCAC
if they are found in the device memory.

+++ src/darwin/Framework/CHIP/CHIPOperationalCredentialsDelegate.h
@@ -67,7 +67,7 @@ private:
     bool ToChipEpochTime(uint32_t offset, uint32_t & epoch);
 
     ChipP256KeypairPtr mIssuerKey;
-    uint32_t mIssuerId = 1234;
+    uint64_t mIssuerId = 1234;
 
     const uint32_t kCertificateValiditySecs = 365 * 24 * 60 * 60;
     const NSString * kCHIPCAKeyChainLabel = @"matter.nodeopcerts.CA:0";
